# swagger/chat.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/chat.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ChatService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/join": {
      "post": {
        "operationId": "ChatService_join",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiJoinResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiUser"
            }
          }
        ],
        "tags": [
          "ChatService"
        ]
      }
    },
    "/v1/sendMsg": {
      "post": {
        "operationId": "ChatService_sendMsg",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiChatMessage"
            }
          }
        ],
        "tags": [
          "ChatService"
        ]
      }
    },
    "/v1/users": {
      "post": {
        "operationId": "ChatService_getAllUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUserList"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ChatService"
        ]
      }
    }
  },
  "definitions": {
    "apiChatMessage": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "time": {
          "type": "string"
        }
      }
    },
    "apiEmpty": {
      "type": "object"
    },
    "apiJoinResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "integer",
          "format": "int32"
        },
        "msg": {
          "type": "string"
        }
      }
    },
    "apiUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "apiUserList": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiUser"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}